Make UploadedFile store fact that it was deleted
We want to make the uploader resilient to changes, and if in some
situations it happens that an already deleted file is attempted to get
deleted again, we would like that no error is raised. This fixed the
delete_invalid plugin, and simplifies moving plugin a bit.

+++ lib/shrine.rb
@@ -745,7 +745,8 @@ class Shrine
 
         # Calls `#delete` on the storage, which deletes the remote file.
         def delete
-          storage.delete(id)
+          storage.delete(id) unless data["deleted"]
+          data["deleted"] = true
         end
 
         # Added as a Ruby conversion method. It typically downloads the file.
